Update to extract the api call for consistency across index pages
