+++ pages/pokemon/alola/index.tsx
@@ -2,8 +2,8 @@ import Layout from '@/components/Layout'
 import PokemonList from '@/components/PokemonList'
 import { GetStaticProps } from 'next';
 import { PokemonIndexPageProps } from '@/interfaces/pokemon';
-import { getPokemonID } from '@/utils/pokemon-utils';
-import { POKE_API, SPRITE_IMAGES } from '@/constants';
+import { getPokemons } from '@/utils/pokemon-utils';
+
 import LinkButton from '@/components/LinkButton';
 
 const WithStaticProps = ({pokemons}: PokemonIndexPageProps) => {
@@ -20,17 +20,9 @@ const WithStaticProps = ({pokemons}: PokemonIndexPageProps) => {
 
 export const getStaticProps: GetStaticProps = async () => {
   try {
-      const response = await POKE_API.listPokemons(721, 88)
-      const results = await response.results
-      const pokemons = results.map((pokemon: { name: string, url: string }) => {
-          const pokemonID = getPokemonID(pokemon.url)
-          const image = `${SPRITE_IMAGES.defaultFront}/${pokemonID}.png`
-          const pokemonPath = `/pokemon/alola/${pokemonID}`
-        return {...pokemon, image, id: pokemonID, pokemonPath }
-      })
-      return {
-        props: { pokemons }
-    }
+    const result = getPokemons({ offset: 721, limit: 88, generationPath: 'pokemon/alola' })
+
+    return result
   } catch (error) {
     console.log(error)
     return {
