+++ pages/pokemon/index.tsx
@@ -2,8 +2,7 @@ import Layout from '@/components/Layout'
 import PokemonList from '@/components/PokemonList'
 import { GetStaticProps } from 'next';
 import { PokemonIndexPageProps } from '@/interfaces/pokemon';
-import { getPokemonID } from '@/utils/pokemon-utils';
-import { POKE_API, SPRITE_IMAGES } from '@/constants';
+import { getPokemons } from '@/utils/pokemon-utils';
 import LinkButton from '@/components/LinkButton';
 
 const WithStaticProps = ({pokemons}: PokemonIndexPageProps) => {
@@ -20,17 +19,9 @@ const WithStaticProps = ({pokemons}: PokemonIndexPageProps) => {
 
 export const getStaticProps: GetStaticProps = async () => {
   try {
-      const response = await POKE_API.listPokemons(0, 1007)
-      const results = await response.results
-      const pokemons = results.map((pokemon: {name: string, url: string}, index: number)=> {
-        const pokemonID = getPokemonID(pokemon.url)
-        const image = `${SPRITE_IMAGES.defaultFront}/${pokemonID}.png`
-        const pokemonPath = `/pokemon/${pokemonID}`
-        return {...pokemon, image, id: pokemonID, pokemonPath }
-      })
-      return {
-        props: { pokemons },
-    }
+    const result = getPokemons({ offset: 0, limit: 1007, generationPath: 'pokemon' })
+
+    return result
   } catch (error) {
     console.log(error)
     return {
