+++ pages/pokemon/johto/index.tsx
@@ -1,7 +1,7 @@
 import Layout from '@/components/Layout'
 import PokemonList from '@/components/PokemonList'
 import { GetStaticProps } from 'next';
-import { getPokemonID } from '@/utils/pokemon-utils';
+import { getPokemonID, getPokemons } from '@/utils/pokemon-utils';
 import { POKE_API, SPRITE_IMAGES } from '@/constants';
 import { PokemonIndexPageProps } from '@/interfaces/pokemon';
 import LinkButton from '@/components/LinkButton';
@@ -20,17 +20,9 @@ const WithStaticProps = ({pokemons}: PokemonIndexPageProps) => {
 
 export const getStaticProps: GetStaticProps = async () => {
   try {
-      const response = await POKE_API.listPokemons(151, 100)
-      const results = await response.results
-      const pokemons = results.map((pokemon: { name: string, url: string }) => {
-          const pokemonID = getPokemonID(pokemon.url)
-          const image = `${SPRITE_IMAGES.defaultFront}/${pokemonID}.png`
-          const pokemonPath = `/pokemon/johto/${pokemonID}`
-        return {...pokemon, image, id: pokemonID, pokemonPath }
-      })
-      return {
-        props: { pokemons }
-    }
+    const result = getPokemons({limit: 100, offset: 151, generationPath: 'pokemon/johto'})
+
+    return result
   } catch (error) {
     console.log(error)
     return {
